Added throwing exceptions

+++ SpringProject/src/main/java/org/example/springproject/services/implementation/AdminServiceImpl.java
@@ -1,9 +1,12 @@
 package org.example.springproject.services.implementation;
 
 import org.example.springproject.entity.Admin;
+import org.example.springproject.exceptions.InvalidNameException;
+import org.example.springproject.exceptions.NoSuchObjectExistsException;
 import org.example.springproject.repository.AdminRepository;
 import org.example.springproject.services.AdminService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
@@ -23,66 +26,65 @@ public class AdminServiceImpl implements AdminService {
 
 	@Override
 	public Admin addAdmin(String name) {
-		try {
-			// Create admin object
-			Admin newAdmin = new Admin();
 
-			// Add new object values
-			newAdmin.setName(name);
-			/*
-			Auto complete the user role, i.e: here we add a new admin
-			 */
-			newAdmin.setRole("admin");
-
-			// Save the object
-			return repository.save(newAdmin);
-
-		} catch (Exception e) {
-			e.printStackTrace();
+		// Verify inserted name
+		for (char c : name.toCharArray()) {
+			if (Character.isDigit(c)) {
+				throw new InvalidNameException("Verify inserted name. Can not use integers in it.", HttpStatus.BAD_REQUEST);
+			}
 		}
 
-		return null;
+		// Create admin object
+		Admin newAdmin = new Admin();
+
+		// Add new object values
+		newAdmin.setName(name);
+		/*
+		Auto complete the user role, i.e: here we add a new admin
+		 */
+		newAdmin.setRole("admin");
+
+		// Save the object
+		return repository.save(newAdmin);
+
 	}
 
 	@Override
 	public Admin updateAdmin(Long id, String name) {
-		try {
-			// Search the admin using id
-			Admin updateAdmin = repository.findAdminById(id);
 
-			// Check if admin not exists
-//			if (updateAdmin == null) {
-//				return ResponseEntity.badRequest().body("Admin with id: " + id + " not found");
-//			}
-
-			// Set the new values of attributes
-			updateAdmin.setName(name);
-			// Save updated object
-			return repository.save(updateAdmin);
-
-		} catch (Exception e) {
-			e.printStackTrace();
+		// Verify inserted name
+		for (char c : name.toCharArray()) {
+			if (Character.isDigit(c)) {
+				throw new InvalidNameException("Verify inserted name. Can not use integers in it.", HttpStatus.BAD_REQUEST);
+			}
 		}
 
-		return null;
+		// Search the admin using id
+		Admin updateAdmin = repository.findAdminById(id);
+
+		 //Check if admin not exists
+		if (updateAdmin == null) {
+			throw new NoSuchObjectExistsException("Admin with id: " + id + " not found.", HttpStatus.NOT_FOUND);
+		}
+
+		// Set the new values of attributes
+		updateAdmin.setName(name);
+		// Save updated object
+		return repository.save(updateAdmin);
+
 	}
 
 	@Override
 	public void deleteAdmin(Long id) {
-		try {
-			// Search wanted admin by id
-			Admin deleteAdmin = repository.findAdminById(id);
+		// Search wanted admin by id
+		Admin deleteAdmin = repository.findAdminById(id);
 
-			// Check if admin not exists
-//			if (deleteAdmin == null) {
-//				return ResponseEntity.badRequest().body("Admin with id: " + id + " not found.");
-//			}
-
-			// Delete found id
-			repository.deleteById(id);
-
-		} catch (Exception e) {
-			e.printStackTrace();
+		// Check if admin not exists
+		if (deleteAdmin == null) {
+			throw new NoSuchObjectExistsException("Admin with id: " + id + " not found.", HttpStatus.NOT_FOUND);
 		}
+
+		// Delete found id
+		repository.deleteById(id);
 	}
 }
